Unify math notation and update structure
