+++ src/construction/signature/pfdh.rs
@@ -13,7 +13,7 @@
 //! works with every PSF and a corresponding hash function.
 
 use super::SignatureScheme;
-use crate::{primitive::hash::HashInto, sample::distribution::psf::PSF};
+use crate::primitive::{hash::HashInto, psf::PSF};
 use qfall_math::{integer::Z, traits::Pow};
 use serde::{Deserialize, Serialize};
 use std::marker::PhantomData;
