+++ src/construction/signature/pfdh/serialize.rs
@@ -9,7 +9,7 @@
 //! Allows to Deserialize an arbitrary [`Pfdh`] instantiation
 
 use super::Pfdh;
-use crate::{primitive::hash::HashInto, sample::distribution::psf::PSF};
+use crate::primitive::{hash::HashInto, psf::PSF};
 use serde::{
     de::{Error, MapAccess, Visitor},
     Deserialize, Serialize,
@@ -119,8 +119,7 @@ where
 mod test_deserialization {
     use crate::{
         construction::signature::{pfdh::Pfdh, SignatureScheme},
-        primitive::hash::HashMatZq,
-        sample::distribution::psf::gpv::PSFGPV,
+        primitive::{hash::HashMatZq, psf::gpv::PSFGPV},
     };
     use qfall_math::{integer::MatZ, integer_mod_q::MatZq, rational::MatQ};
