--- src/sample/distribution/discrete_gauss.rs
@@ -1,1 +1,0 @@
-
