--- src/sample/distribution/uniform.rs
@@ -1,1 +1,0 @@
-
