--- src/sample/lattice_problem.rs
@@ -1,13 +1,0 @@
-// Copyright © 2023 Niklas Siemer
-//
-// This file is part of qFALL-crypto.
-//
-// qFALL-crypto is free software: you can redistribute it and/or modify it under
-// the terms of the Mozilla Public License Version 2.0 as published by the
-// Mozilla Foundation. See <https://mozilla.org/en-US/MPL/2.0/>.
-
-//! This module provides the possibility to quickly sample instances of lattice
-//! problems like `LWE`, `SIS`, and more.
-
-pub mod lwe;
-pub mod sis;
